My Replication of Dupuy and Galichon "Personality Traits and the Marriage Market" (2014)

@Gabriele Buontempo @mattjs07

This replication study was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we do translate the original matlab code into Julia lang.

TESTING IF CHANGES ARE DONE.

Dupuy_Galichon_Matching.convertion — Method

convertion(x::Any)

Converts element to Float64, usefull to convert String. It is especially coded to convert the strings using commas.
convertion("0,01") –> returns Float64(0.01)

Dupuy_Galichon_Matching.convertion_df! — Method

convertion_df(x::DataFrame)

Converts every elements of a DataFrame to type Float64.
Based on convertion()

Dupuy_Galichon_Matching.z_diag — Method

z_diag(x::Vector)

Returns a zero matrix with the input vector as diagonal.

Dupuy_Galichon_Matching.z_diag_f — Method

z_diag_f(x::Matrix, f::Function)

Mimick the matlab function diag().
Apply a function on a Matrix columns (such as mean() or var()).
Returns a zero matrix with the computed values in its diagonal.

end