Our Replication of Dupuy and Galichon "Personality Traits and the Marriage Market" (2014)

@Gabriele Buontempo @mattjs07

This replication study was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we translate the original matlab code into Julia lang. You will find in this documentation the main function " AffinityMatrix(df::DataFrame)" as well as the intermediary functions we built for this function. Unfortunately, we were not able to achieve the last step of optimisation, thereby the AffinityMatrix() function is not operational.

Dupuy_Galichon_Matching.AffinityMatrix — Method

AffinityMatrix(df::DataFrame) Function allowing to compute the Affinity Matrix between two groups X and Y.
For m characteristics per group, returns a m x m Affinity Matrix.
!!! The input Dataframe contains columns of characteristics which must be ordered as follow :
> The Number of characteristics for X must be equal to the number of characteristics of Y.
> If the total number of columns = n, then columns 1:n/2 are X's and columns n/2+1:n are Y's.
> In each group the column must be ordered alike.
> e.g A Dataframe with 3 characteristics, the columns are arranged as : [X.education, X.age, X.sex , Y.education, Y.age, Y.sex ]

Dupuy_Galichon_Matching.CovVarEdsi — Method

CovVarEdsi(X::Matrix,Y::Matrix )

Returns Cov _XY _XY , Cov _XX _XX, Cov _YY _YY, Cov _XX _YY

Dupuy_Galichon_Matching.EmpCov — Method

EmpCov(X::Matrix,Y::Matrix)

Returns the empirical covariance of two matrices.

Dupuy_Galichon_Matching.ObjectiveFunction — Method

ObjectiveFunction(X::Matrix,Y::Matrix,sigmaHat::Matrix,T::Int64,A... )

Objective function used to compute the optimal Affinity Matrix.
A is splat as it is the argument over which we arg min using JuMP.

Dupuy_Galichon_Matching.OptimalPi — Method

OptimalPi(auxVar::Matrix, X::Matrix, Y::Matrix, T::Int64)

Returns the Optimal Pi, used in ObjectiveFunction() for optimization.

Dupuy_Galichon_Matching.StartingValues — Method

StartingValues(X::Matrix, Y::Matrix)  

N,M = size(X)
u0 = log(N) * ones(N,1)
v0 = -ones(N,1)
return u0, v0

Used in OptimalPi() function.

Dupuy_Galichon_Matching.convertion — Method

convertion(x::Any)

Converts element to Float64, usefull to convert String. It is especially coded to convert the strings using commas.
convertion("0,01") –> returns Float64(0.01)

Dupuy_Galichon_Matching.convertion_df! — Method

convertion_df(x::DataFrame)

Converts every elements of a DataFrame to type Float64.
Based on convertion()

Dupuy_Galichon_Matching.discretePdf — Method

discretePdf(X::Matrix, Y::Matrix)

Returns two vectors of length equal to the matrices highest dimension (N).
Each element equals to 1 divided by the number of rows. Used in OptimalPi() function.

Dupuy_Galichon_Matching.length_ — Method

length_(X::Matrix)

Mimick matlab length() function by returning the length of the biggest dimension.

Dupuy_Galichon_Matching.mm — Method

mm(x::Matrix, y::Matrix)

Performs matrice multiplication using loops.
Usefull for optimisation with JuMP.

Dupuy_Galichon_Matching.xlogx — Method

xlogx(x::Matrix)
return  x .* log.(x + eps() * (x .<= 0))

Dupuy_Galichon_Matching.z_diag — Method

z_diag(x::Vector)

Returns a zero matrix with the input vector as diagonal.

Dupuy_Galichon_Matching.z_diag_f — Method

z_diag_f(x::Matrix, f::Function)

Mimick the matlab function diag().
Apply a function on a Matrix columns (such as mean() or var()).
Returns a zero matrix with the computed values in its diagonal.

end